Revert "chore: apply each block controlled teardown optimization (#11045)"
This reverts commit 1afec80261f09551b8daa664bd946f25c7a8b48a.
